Moved code and config to slp2mp4 directory, factored out config module
