+++ slp2mp4/User/Config/Dolphin.ini
@@ -18,7 +18,7 @@ AutoHideCursor = False
 MainWindowPosX = 0
 MainWindowPosY = 495
 MainWindowWidth = 1366
-MainWindowHeight = 1145
+MainWindowHeight = 1131
 LanguageCode = 
 ShowToolbar = False
 ShowStatusbar = False
