+++ slp2mp4/User/GameSettings/GALE01.ini
