+++ slp2mp4/User/GameSettings/PALE02.ini
